requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

psycopg2-binary==2.9.9

pika==1.3.2

python-decouple==3.8
drf-yasg==1.21.7
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

python-slugify==8.0.1

django-prometheus==2.3.1

pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

gunicorn==21.2.0
whitenoise==6.6.0